tests/wpt/tests/html/dom/render-blocking/element-render-blocking-016.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/utils.js"></script>
<title>Removing blocking attr in the body has an effect</title>

<link id=link rel=expect href="#last" blocking="render">
<script>
async_test((t) => {
  requestAnimationFrame(() => {
    t.step(() => assert_false(!!document.getElementById("last")));
    t.done();
  });
}, "removing 'blocking' in the body makes it non blocking");
</script>
</head>
<body>
<script>
link.blocking = "";
</script>

  <div id="first"></div>
  <script>
          generateParserDelay();
  </script>
  <div id="second"></div>
  <script>
          generateParserDelay();
  </script>
  <div id="last"></div>
</body>
